@@ -16,7 +16,7 @@ class Inscriptions
      * @ORM\GeneratedValue
      * @ORM\Column(type="integer")
      */
-    private ?int $id;
+    private ?int $id = null;
 
     /**
      * @ORM\Column(type="string", length=255, nullable=true)
@@ -31,32 +31,10 @@ class Inscriptions
     private ?string $lastname;
 
     /**
-     * @ORM\Column(type="boolean", nullable=true)
-     */
-    private $isOk = null;
-
-    /**
      * @ORM\Column(type="string", length=255, nullable=true, unique=true)
      * @Assert\NotBlank()
      */
     private ?string $email;
-
-    /**
-     * @ORM\Column(type="integer")
-     * @Assert\NotBlank()
-     */
-    private ?int $phone;
-
-    /**
-     * @ORM\Column(type="string", length=255, nullable=true)
-     */
-    private ?string $adress;
-
-    /**
-     * @ORM\Column(type="integer", nullable=true)
-     * @Assert\NotBlank()
-     */
-    private ?int $cp;
 
     /**
      * @ORM\Column(type="string", length=255, nullable=true)
@@ -65,17 +43,9 @@ class Inscriptions
     private ?string $city;
 
     /**
-     * @ORM\Column(type="string", length=255, nullable=true)
-     * @Assert\NotBlank()
-     */
-    private ?string $department;
-
-    /**
      * @ORM\Column(type="text", nullable=true)
      */
     private ?string $note;
-
-    
 
 
     public function getId(): ?int
@@ -83,11 +53,11 @@ class Inscriptions
         return $this->id;
     }
 
+
     public function getFirstname(): ?string
     {
         return $this->firstname;
     }
-
     public function setFirstname(?string $firstname): self
     {
         $this->firstname = $firstname;
@@ -95,11 +65,11 @@ class Inscriptions
         return $this;
     }
 
+
     public function getLastname(): ?string
     {
         return $this->lastname;
     }
-
     public function setLastname(?string $lastname): self
     {
         $this->lastname = $lastname;
@@ -107,23 +77,11 @@ class Inscriptions
         return $this;
     }
 
-    public function getIsOk(): ?bool
-    {
-        return $this->isOk;
-    }
-
-    public function setIsOk(?bool $isOk): self
-    {
-        $this->isOk = $isOk;
-
-        return $this;
-    }
 
     public function getEmail(): ?string
     {
         return $this->email;
     }
-
     public function setEmail(?string $email): self
     {
         $this->email = $email;
@@ -131,47 +89,11 @@ class Inscriptions
         return $this;
     }
 
-    public function getPhone(): ?int
-    {
-        return $this->phone;
-    }
-
-    public function setPhone(?int $phone): self
-    {
-        $this->phone = $phone;
-
-        return $this;
-    }
-
-    public function getAdress(): ?string
-    {
-        return $this->adress;
-    }
-
-    public function setAdress(?string $adress): self
-    {
-        $this->adress = $adress;
-
-        return $this;
-    }
-
-    public function getCp(): ?int
-    {
-        return $this->cp;
-    }
-
-    public function setCp(?int $cp): self
-    {
-        $this->cp = $cp;
-
-        return $this;
-    }
 
     public function getCity(): ?string
     {
         return $this->city;
     }
-
     public function setCity(?string $city): self
     {
         $this->city = $city;
@@ -179,23 +101,11 @@ class Inscriptions
         return $this;
     }
 
-    public function getDepartment(): ?string
-    {
-        return $this->department;
-    }
-
-    public function setDepartment(?string $department): self
-    {
-        $this->department = $department;
-
-        return $this;
-    }
 
     public function getNote(): ?string
     {
         return $this->note;
     }
-
     public function setNote(?string $note): self
     {
         $this->note = $note;
